requirements.txt
flask
python-dotenv
supabase
psycopg2-binary
PyMuPDF
sentence-transformers
groq
gunicorn
